+++ src/main/java/bowling/domain/concrete/frame/NormalFrame.java
@@ -1,21 +1,20 @@
-package bowling.domain.concrete;
+package bowling.domain.concrete.frame;
 
 import bowling.domain.engine.frame.Frame;
 import bowling.domain.engine.frame.state.State;
 import bowling.domain.engine.frame.state.StateFactory;
 import bowling.domain.engine.roll.RollResult;
 
-
-public class FinalFrame implements Frame {
+public class NormalFrame implements Frame {
 
     private State state;
 
-    private FinalFrame(State state) {
+    private NormalFrame(State state) {
         this.state = state;
     }
 
-    public static FinalFrame init() {
-        return new FinalFrame(StateFactory.ready());
+    public static NormalFrame init() {
+        return new NormalFrame(StateFactory.ready());
     }
 
     @Override
@@ -24,17 +23,12 @@ public class FinalFrame implements Frame {
             throw new IllegalStateException("이미 프레임이 종료된 상태입니다.");
         }
 
-        if (state.isFinished() && state.canPromoteToBonusState()) {
-            state = state.continueInBonus().transit(rollResult);
-            return ;
-        }
-
-        state = state.transit(rollResult);
+        this.state = state.transit(rollResult);
     }
 
     @Override
     public boolean isEnded() {
-        return state.isFinished() && !state.canPromoteToBonusState();
+        return state.isFinished();
     }
 
     @Override
