+++ src/main/java/bowling/domain/concrete/roll/NormalResult.java
@@ -1,0 +1,11 @@
+package bowling.domain.concrete.roll;
+
+import bowling.domain.engine.roll.RollResult;
+
+public class NormalResult extends RollResult {
+
+    public NormalResult(int value) {
+        super(value);
+    }
+
+}
