--- src/main/java/bowling/domain/engine/roll/NormalResult.java
@@ -1,9 +1,0 @@
-package bowling.domain.engine.roll;
-
-class NormalResult extends RollResult {
-
-    NormalResult(int value) {
-        super(value);
-    }
-
-}
