+++ src/test/java/bowling/domain/engine/frame/FrameCreatorTest.java
@@ -1,7 +1,7 @@
 package bowling.domain.engine.frame;
 
-import bowling.domain.concrete.FinalFrame;
-import bowling.domain.concrete.NormalFrame;
+import bowling.domain.concrete.frame.FinalFrame;
+import bowling.domain.concrete.frame.NormalFrame;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
