+++ src/test/java/bowling/domain/engine/roll/RollTest.java
@@ -1,5 +1,7 @@
 package bowling.domain.engine.roll;
 
+import bowling.domain.concrete.roll.GutterResult;
+import bowling.domain.concrete.roll.NormalResult;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
